requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
attrs==17.4.0
autobahn==18.4.1
Automat==0.6.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.2.2
dateparser==0.7.0
eventlet==0.23.0
Events==0.3
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==0.4.13
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
passlib==1.7.1
playsound==1.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-binance-api==1.1.0
python-dateutil==2.7.2
python-editor==1.0.3
python-engineio==2.1.1
python-socketio==1.9.0
pytz==2018.4
regex==2018.2.21
requests==2.18.4
service-identity==17.0.0
simplejson==3.14.0
six==1.11.0
socketIO-client==0.7.2
SQLAlchemy==1.2.9
Twisted==18.4.0
txaio==2.10.0
tzlocal==1.5.1
urllib3==1.22
websocket-client==0.47.0
websockets==5.0.1
Werkzeug==0.14.1
WTForms==2.2.1
zope.interface==4.5.0